Rename package to Swift Locking
