+++ .github/ISSUE_TEMPLATE/question.yml
@@ -10,7 +10,7 @@ body:
       value: |
         ### Thanks for asking a question!
         > [!Important]
-        > Please consider using [Discussions](https://github.com/fetch-rewards/swift-synchronization/discussions) to ask your question. There, you can receive answers from maintainers as well as community members. While we do provide this form for submitting questions, we prefer to reserve issues for things that need to be fixed, changed, or added. If you are sharing feedback, asking for help or opinions, or looking to brainstorm ideas with other members of the community, please use [Discussions](https://github.com/fetch-rewards/swift-synchronization/discussions).
+        > Please consider using [Discussions](https://github.com/fetch-rewards/swift-locking/discussions) to ask your question. There, you can receive answers from maintainers as well as community members. While we do provide this form for submitting questions, we prefer to reserve issues for things that need to be fixed, changed, or added. If you are sharing feedback, asking for help or opinions, or looking to brainstorm ideas with other members of the community, please use [Discussions](https://github.com/fetch-rewards/swift-locking/discussions).
         
   - type: textarea
     id: question
@@ -25,7 +25,7 @@ body:
     attributes:
       label: Checklist
       options:
-        - label: I have searched existing [Issues](https://github.com/fetch-rewards/swift-synchronization/issues) and [Discussions](https://github.com/fetch-rewards/swift-synchronization/discussions).
+        - label: I have searched existing [Issues](https://github.com/fetch-rewards/swift-locking/issues) and [Discussions](https://github.com/fetch-rewards/swift-locking/discussions).
           required: true
         - label: This is not a bug report or feature request.
           required: true
@@ -34,7 +34,7 @@ body:
     id: terms
     attributes:
       label: Code of Conduct
-      description: By submitting this issue, you agree to follow our [Code of Conduct](https://github.com/fetch-rewards/swift-synchronization/blob/main/CODE_OF_CONDUCT.md). 
+      description: By submitting this issue, you agree to follow our [Code of Conduct](https://github.com/fetch-rewards/swift-locking/blob/main/CODE_OF_CONDUCT.md). 
       options:
         - label: I agree to follow this project's Code of Conduct
           required: true
