+++ .github/workflows/ci.yml
@@ -44,13 +44,13 @@ jobs:
         run: swift test --enable-code-coverage
 
       - name: Gather code coverage
-        run: xcrun llvm-cov export -format="lcov" .build/debug/swift-synchronizationPackageTests.xctest/Contents/MacOS/swift-synchronizationPackageTests -instr-profile .build/debug/codecov/default.profdata > coverage_report.lcov
+        run: xcrun llvm-cov export -format="lcov" .build/debug/swift-lockingPackageTests.xctest/Contents/MacOS/swift-lockingPackageTests -instr-profile .build/debug/codecov/default.profdata > coverage_report.lcov
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v5
         with:
           files: coverage_report.lcov
           token: ${{ secrets.CODECOV_TOKEN }}
-          slug: fetch-rewards/swift-synchronization
+          slug: fetch-rewards/swift-locking
           verbose: true
           fail_ci_if_error: true
