+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org).
 
-## 📦 [Version 0.1.0](https://github.com/fetch-rewards/swift-synchronization/releases/tag/0.1.0) - April 23, 2025 ([Full Changelog](https://github.com/fetch-rewards/swift-synchronization/commits/0.1.0))
+## 📦 [Version 0.1.0](https://github.com/fetch-rewards/swift-locking/releases/tag/0.1.0) - April 23, 2025 ([Full Changelog](https://github.com/fetch-rewards/swift-locking/commits/0.1.0))
 
 ### 🚀 Initial Release
 
-This is the first public release of Swift Synchronization, a library that provides a collection of Swift macros used to protect shared mutable state.
+This is the first public release of Swift Locking, a library that provides a collection of Swift macros used to protect shared mutable state.
 
 This initial release includes:
