+++ Sources/Locking/Macros/Locked/Locked.swift
@@ -9,6 +9,6 @@
 @attached(accessor, names: named(init), named(get), named(set))
 @attached(peer, names: prefixed(_))
 public macro Locked(_ lockType: LockType) = #externalMacro(
-    module: "SynchronizationMacros",
+    module: "LockingMacros",
     type: "LockedMacro"
 )
