+++ Sources/LockingMacros/LockingPlugin.swift
@@ -1,5 +1,5 @@
 //
-//  SynchronizationPlugin.swift
+//  LockingPlugin.swift
 //
 //  Copyright © 2025 Fetch.
 //
@@ -8,7 +8,7 @@ import SwiftCompilerPlugin
 import SwiftSyntaxMacros
 
 @main
-struct SynchronizationPlugin: CompilerPlugin {
+struct LockingPlugin: CompilerPlugin {
     let providingMacros: [any Macro.Type] = [
         LockedMacro.self,
     ]
