+++ Sources/LockingMacros/Macros/LockedMacro/LockedMacro+MacroError.swift
