+++ Sources/LockingMacros/Models/LockType/LockType.swift
