+++ Tests/LockingMacrosTests/Macros/LockedMacro/TestHelpers/AssertMacroExpansion.swift
@@ -7,11 +7,11 @@
 // Macro implementations build for the host, so the corresponding module is not
 // available when cross-compiling. Cross-compiled tests may still make use of
 // the macro itself in end-to-end tests.
-#if canImport(SynchronizationMacros)
+#if canImport(LockingMacros)
 import SwiftSyntaxMacroExpansion
 import SwiftSyntaxMacrosGenericTestSupport
 import Testing
-@testable import SynchronizationMacros
+@testable import LockingMacros
 
 func assertMacroExpansion(
     _ originalSource: String,
